Extract `PrimitiveType` extensions
