+++ java/src/main/kotlin/io/spine/validation/java/PrimitiveTypeExtensions.kt
@@ -1,0 +1,82 @@
+/*
+ * Copyright 2022, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.validation.java
+
+import com.google.protobuf.ByteString
+import io.spine.protodata.PrimitiveType
+import io.spine.protodata.PrimitiveType.PT_UNKNOWN
+import io.spine.protodata.PrimitiveType.TYPE_BOOL
+import io.spine.protodata.PrimitiveType.TYPE_BYTES
+import io.spine.protodata.PrimitiveType.TYPE_DOUBLE
+import io.spine.protodata.PrimitiveType.TYPE_FIXED32
+import io.spine.protodata.PrimitiveType.TYPE_FIXED64
+import io.spine.protodata.PrimitiveType.TYPE_FLOAT
+import io.spine.protodata.PrimitiveType.TYPE_INT32
+import io.spine.protodata.PrimitiveType.TYPE_INT64
+import io.spine.protodata.PrimitiveType.TYPE_SFIXED32
+import io.spine.protodata.PrimitiveType.TYPE_SFIXED64
+import io.spine.protodata.PrimitiveType.TYPE_SINT32
+import io.spine.protodata.PrimitiveType.TYPE_SINT64
+import io.spine.protodata.PrimitiveType.TYPE_STRING
+import io.spine.protodata.PrimitiveType.TYPE_UINT32
+import io.spine.protodata.PrimitiveType.TYPE_UINT64
+import io.spine.protodata.PrimitiveType.UNRECOGNIZED
+import io.spine.protodata.codegen.java.ClassName
+import kotlin.reflect.KClass
+
+/**
+ * Obtains a name of the class which corresponds to this primitive type.
+ */
+internal fun PrimitiveType.toClass(): ClassName {
+    val klass = primitiveClass()
+    return ClassName(klass.javaObjectType)
+}
+
+/**
+ * Obtains a name of the class which corresponds to this primitive type.
+ */
+internal fun PrimitiveType.toPrimitiveName(): String {
+    val klass = primitiveClass()
+    val primitiveClass = klass.javaPrimitiveType ?: klass.java
+    return primitiveClass.name
+}
+
+private fun PrimitiveType.primitiveClass(): KClass<*> =
+    when (this) {
+        TYPE_DOUBLE -> Double::class
+        TYPE_FLOAT -> Float::class
+        TYPE_INT64, TYPE_UINT64, TYPE_SINT64, TYPE_FIXED64, TYPE_SFIXED64 -> Long::class
+        TYPE_INT32, TYPE_UINT32, TYPE_SINT32, TYPE_FIXED32, TYPE_SFIXED32 -> Int::class
+        TYPE_BOOL -> Boolean::class
+        TYPE_STRING -> String::class
+        TYPE_BYTES -> ByteString::class
+        UNRECOGNIZED, PT_UNKNOWN -> unknownType(this)
+    }
+
+private fun unknownType(type: PrimitiveType): Nothing {
+    throw IllegalStateException("Unknown primitive type: `$type`.")
+}
